Used InventoryTracker, deleted PartInventory.cs
