--- Assets/Scripts/Control/PartInventory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 64b2ab43392586b439d9a436c2be4e94
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
