+++ SpaceBuilder.sln
@@ -1,9 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31112.23
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{38E24355-6D1D-7254-5CF4-D160E2B43788}"
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{41A5F2E1-B838-6191-8952-979F4D4E9F6E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +9,12 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{38E24355-6D1D-7254-5CF4-D160E2B43788}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{38E24355-6D1D-7254-5CF4-D160E2B43788}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{38E24355-6D1D-7254-5CF4-D160E2B43788}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{38E24355-6D1D-7254-5CF4-D160E2B43788}.Release|Any CPU.Build.0 = Release|Any CPU
+		{41A5F2E1-B838-6191-8952-979F4D4E9F6E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{41A5F2E1-B838-6191-8952-979F4D4E9F6E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{41A5F2E1-B838-6191-8952-979F4D4E9F6E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{41A5F2E1-B838-6191-8952-979F4D4E9F6E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4809A3EB-21D4-47C9-88A4-7A8F388BF903}
 	EndGlobalSection
 EndGlobal
